python/requirements.txt
pynput
todoist-python
